refactor: RequestParser의 역할 분리
RequestParser가 요청을 파싱하는 역할과 HttpRequest 정보를 가지고 있는 역할 중 파싱하는 역할만 하도록 수정.
HttpRequest 클래스 추가 -> HttpRequest 정보를 가지고 있는 역할
