+++ src/main/java/util/RequestParser.java
@@ -9,40 +9,19 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
 public class RequestParser {
     private static final Logger logger = LoggerFactory.getLogger(RequestParser.class);
 
-    private List<String> headers;
-    private String method;
-    private String target;
-    private final Map<String,String> queries = new HashMap<>();
-    private String version;
-
     public RequestParser() {}
 
-    public void parse(InputStream in) throws IOException, InvalidRequestLineSyntaxException{
-        headers = logAndReturnHeaders(in);
+    public HttpRequest parse(InputStream in) throws IOException, InvalidRequestLineSyntaxException{
+        List<String> headers = logAndReturnHeaders(in);
 
         String[] requestLine = resolveRequestLine(headers.get(0));
-        method = requestLine[0];
-        version = requestLine[2];
 
-        String[] requestTarget = resolveRequestTarget(requestLine[1]);
-
-        target = requestTarget[0];
-        if (requestTarget.length > 1) {
-            String[] queryArray = resolveQuery(requestTarget[1]);
-            for (String s : queryArray) {
-                String[] items = s.split("=");
-                String key = items[0];
-                String value = items.length > 1 ? items[1] : null;
-                queries.put(key, value);
-            }
-        }
+        return new HttpRequest(requestLine[0], requestLine[1], requestLine[2], headers.subList(1, headers.size()));
     }
 
     private List<String> logAndReturnHeaders(InputStream in) throws IOException {
@@ -69,33 +48,5 @@ public class RequestParser {
         }
 
         return tokens;
-    }
-
-    private String[] resolveRequestTarget(String target) {
-        return target.split("\\?");
-    }
-
-    private String[] resolveQuery(String query) {
-        return query.split("&");
-    }
-
-    public List<String> getHeaders() {
-        return headers;
-    }
-
-    public String getMethod() {
-        return method;
-    }
-
-    public String getTarget() {
-        return target;
-    }
-
-    public Map<String, String> getQueries() {
-        return queries;
-    }
-
-    public String getVersion() {
-        return version;
     }
 }
